+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_options.hpp
@@ -6,6 +6,7 @@
 #include <chrono>
 #include <cstdint>
 #include <memory>
+#include <numeric>
 #include <string>
 #include <vector>
 
@@ -13,6 +14,7 @@
 #include <azure/core/match_conditions.hpp>
 #include <azure/core/modified_conditions.hpp>
 #include <azure/storage/common/access_conditions.hpp>
+#include <azure/storage/common/crypt.hpp>
 
 #include "azure/storage/blobs/rest_client.hpp"
 
@@ -155,7 +157,7 @@ namespace Azure { namespace Storage { namespace Blobs {
     /**
      * API version used by this client.
      */
-    std::string ApiVersion = _detail::ApiVersion;
+    std::string ApiVersion;
   };
 
   /**
@@ -289,10 +291,22 @@ namespace Azure { namespace Storage { namespace Blobs {
   };
 
   /**
-   * @brief Optional parameters for #Azure::Storage::Blobs::BlobContainerClient::Undelete.
+   * @brief Optional parameters for
+   * #Azure::Storage::Blobs::BlobServiceClient::UndeleteBlobContainer.
    */
   struct UndeleteBlobContainerOptions final
   {
+  };
+
+  /**
+   * @brief Optional parameters for #Azure::Storage::Blobs::BlobServiceClient::RenameBlobContainer.
+   */
+  struct RenameBlobContainerOptions final
+  {
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    LeaseAccessConditions SourceAccessConditions;
   };
 
   /**
@@ -495,6 +509,71 @@ namespace Azure { namespace Storage { namespace Blobs {
      * @brief If the destination blob should be sealed. Only applicable for Append Blobs.
      */
     Azure::Nullable<bool> ShouldSealDestination;
+
+    /**
+     * Immutability policy to set on the destination blob.
+     */
+    Azure::Nullable<Models::BlobImmutabilityPolicy> ImmutabilityPolicy;
+
+    /**
+     * Indicates whether the destination blob has a legal hold.
+     */
+    Azure::Nullable<bool> HasLegalHold;
+  };
+
+  /**
+   * @brief Optional parameters for #Azure::Storage::Blobs::BlobClient::CopyFromUri.
+   */
+  struct CopyBlobFromUriOptions
+  {
+    /**
+     * @brief Specifies user-defined name-value pairs associated with the blob. If no
+     * name-value pairs are specified, the operation will copy the metadata from the source blob or
+     * file to the destination blob. If one or more name-value pairs are specified, the destination
+     * blob is created with the specified metadata, and metadata is not copied from the source blob
+     * or file.
+     */
+    Storage::Metadata Metadata;
+
+    /**
+     * @brief The tags to set for this blob.
+     */
+    std::map<std::string, std::string> Tags;
+
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    BlobAccessConditions AccessConditions;
+
+    /**
+     * @brief Optional conditions that the source must meet to perform this operation.
+     *
+     * @note Lease access condition only works for API versions before 2012-02-12.
+     */
+    struct : public Azure::ModifiedConditions, public Azure::MatchConditions
+    {
+    } SourceAccessConditions;
+
+    /**
+     * @brief Specifies the tier to be set on the target blob.
+     */
+    Azure::Nullable<Models::AccessTier> AccessTier;
+
+    /**
+     * @brief Hash of the blob content. This hash is used to verify the integrity of
+     * the blob during transport. When this header is specified, the storage service checks the hash
+     * that has arrived with the one that was sent.
+     */
+    Azure::Nullable<ContentHash> TransactionalContentHash;
+
+    /**
+     * Immutability policy to set on the destination blob.
+     */
+    Azure::Nullable<Models::BlobImmutabilityPolicy> ImmutabilityPolicy;
+    /**
+     * Indicates whether the destination blob has a legal hold.
+     */
+    Azure::Nullable<bool> HasLegalHold;
   };
 
   /**
@@ -680,7 +759,9 @@ namespace Azure { namespace Storage { namespace Blobs {
     /**
      * @brief Optional conditions that must be met to perform this operation.
      */
-    TagAccessConditions AccessConditions;
+    struct : public LeaseAccessConditions, public TagAccessConditions
+    {
+    } AccessConditions;
   };
 
   /**
@@ -691,7 +772,9 @@ namespace Azure { namespace Storage { namespace Blobs {
     /**
      * @brief Optional conditions that must be met to perform this operation.
      */
-    TagAccessConditions AccessConditions;
+    struct : public LeaseAccessConditions, public TagAccessConditions
+    {
+    } AccessConditions;
   };
 
   /**
@@ -730,6 +813,16 @@ namespace Azure { namespace Storage { namespace Blobs {
      * @brief Optional conditions that must be met to perform this operation.
      */
     BlobAccessConditions AccessConditions;
+
+    /**
+     * Immutability policy to set on the blob.
+     */
+    Azure::Nullable<Models::BlobImmutabilityPolicy> ImmutabilityPolicy;
+
+    /**
+     * Indicates whether the blob has a legal hold.
+     */
+    Azure::Nullable<bool> HasLegalHold;
   };
 
   /**
@@ -779,6 +872,66 @@ namespace Azure { namespace Storage { namespace Blobs {
        */
       int32_t Concurrency = 5;
     } TransferOptions;
+
+    /**
+     * Immutability policy to set on the blob.
+     */
+    Azure::Nullable<Models::BlobImmutabilityPolicy> ImmutabilityPolicy;
+
+    /**
+     * Indicates whether the blob has a legal hold.
+     */
+    Azure::Nullable<bool> HasLegalHold;
+  };
+
+  struct UploadBlockBlobFromUriOptions final
+  {
+    /**
+     * If true, the properties of the source blob will be copied to the new blob.
+     */
+    bool CopySourceBlobProperties = true;
+
+    /**
+     * @brief The standard HTTP header system properties to set.
+     */
+    Models::BlobHttpHeaders HttpHeaders;
+
+    /**
+     * @brief Name-value pairs associated with the blob as metadata.
+     */
+    Storage::Metadata Metadata;
+
+    /**
+     * @brief The tags to set for this blob.
+     */
+    std::map<std::string, std::string> Tags;
+
+    /**
+     * @brief Indicates the tier to be set on blob.
+     */
+    Azure::Nullable<Models::AccessTier> AccessTier;
+
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    BlobAccessConditions AccessConditions;
+
+    /**
+     * @brief Optional conditions that source must meet to perform this operation.
+     */
+    struct : public Azure::ModifiedConditions,
+             public Azure::MatchConditions,
+             public TagAccessConditions
+    {
+    } SourceAccessConditions;
+
+    /**
+     * @brief Hash of the blob content. This hash is used to verify the integrity of
+     * the blob during transport. When this header is specified, the storage service checks the hash
+     * that has arrived with the one that was sent. Note that this hash is not stored with the blob.
+     * If the two hashes do not match, the operation will fail.
+     */
+    Azure::Nullable<ContentHash> TransactionalContentHash;
   };
 
   /**
@@ -858,6 +1011,16 @@ namespace Azure { namespace Storage { namespace Blobs {
      * @brief Optional conditions that must be met to perform this operation.
      */
     BlobAccessConditions AccessConditions;
+
+    /**
+     * Immutability policy to set on the blob.
+     */
+    Azure::Nullable<Models::BlobImmutabilityPolicy> ImmutabilityPolicy;
+
+    /**
+     * Indicates whether the blob has a legal hold.
+     */
+    Azure::Nullable<bool> HasLegalHold;
   };
 
   /**
@@ -903,6 +1066,16 @@ namespace Azure { namespace Storage { namespace Blobs {
      * @brief Optional conditions that must be met to perform this operation.
      */
     BlobAccessConditions AccessConditions;
+
+    /**
+     * Immutability policy to set on the blob.
+     */
+    Azure::Nullable<Models::BlobImmutabilityPolicy> ImmutabilityPolicy;
+
+    /**
+     * Indicates whether the blob has a legal hold.
+     */
+    Azure::Nullable<bool> HasLegalHold;
   };
 
   /**
@@ -994,6 +1167,16 @@ namespace Azure { namespace Storage { namespace Blobs {
      * @brief Optional conditions that must be met to perform this operation.
      */
     BlobAccessConditions AccessConditions;
+
+    /**
+     * Immutability policy to set on the blob.
+     */
+    Azure::Nullable<Models::BlobImmutabilityPolicy> ImmutabilityPolicy;
+
+    /**
+     * Indicates whether the blob has a legal hold.
+     */
+    Azure::Nullable<bool> HasLegalHold;
   };
 
   /**
@@ -1105,4 +1288,49 @@ namespace Azure { namespace Storage { namespace Blobs {
     BlobAccessConditions AccessConditions;
   };
 
+  /**
+   * @brief Optional parameters for #Azure::Storage::Blobs::BlobClient::SetLegalHold.
+   */
+  struct SetBlobLegalHoldOptions final
+  {
+  };
+
+  /**
+   * @brief Optional parameters for #Azure::Storage::Blobs::BlobClient::SetImmutabilityPolicy.
+   */
+  struct SetBlobImmutabilityPolicyOptions final
+  {
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    struct
+    {
+      /**
+       * @brief Specify this header to perform the operation only if the resource has not been
+       * modified since the specified time. This timestamp will be truncated to second.
+       */
+      Azure::Nullable<Azure::DateTime> IfUnmodifiedSince;
+    } AccessConditions;
+  };
+
+  /**
+   * @brief Optional parameters for #Azure::Storage::Blobs::BlobClient::DeleteImmutabilityPolicy.
+   */
+  struct DeleteBlobImmutabilityPolicyOptions final
+  {
+  };
+
+  namespace _detail {
+    inline std::string TagsToString(const std::map<std::string, std::string>& tags)
+    {
+      return std::accumulate(
+          tags.begin(),
+          tags.end(),
+          std::string(),
+          [](const std::string& a, const std::pair<std::string, std::string>& b) {
+            return a + (a.empty() ? "" : "&") + _internal::UrlEncodeQueryParameter(b.first) + "="
+                + _internal::UrlEncodeQueryParameter(b.second);
+          });
+    }
+  } // namespace _detail
 }}} // namespace Azure::Storage::Blobs