+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_options.hpp
@@ -43,7 +43,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     /**
      * API version used by this client.
      */
-    std::string ApiVersion = _detail::ApiVersion;
+    std::string ApiVersion;
   };
 
   /**